requirements.txt
discord.py==2.5.1
python-dotenv==1.0.0
aiohttp==3.8.1
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.80.0
facebook-business==16.0.0

